requirements.txt
click==8.3.1
langchain==1.2.7
langchain_community==0.4.1
langchain_core==1.2.7
langchain_openai==1.1.7
langchain_text_splitters==1.1.0
langgraph==1.0.7
python-dotenv==1.2.1
questionary==2.1.1
rich==14.3.1
faiss-cpu==1.13.2
